Changed to much for one line check desc
Reworked the selection of tickets available > Redid the innerjoin now it selects the departure and destination(How foolish to forget that...)
Created more tests.
Created an error for not selecting a booking request in overview
Created exception for ticket already sold. (if it would happen that someone for somereason could do that.)